Fixed So Many Things!
Campfire is now showing the correct $pagetitle & h3. Pages are going to
the right url (excluding cfstories and tc, will work on that next).

Need to work on changing where the forms upload to, want to have them
go to a local PHP file. Also, need to fix TC and Portfolio arrays.

*Delete img resources I no longer need and upload ones I do.